+++ src/pycropml/checking.py
@@ -1,0 +1,8 @@
+
+class Test(object):
+    """ Test """
+    def __init__(self, name, description, uri=None):
+        self.name = name
+        self.description = description
+        self.uri = uri
+        self.paramsets = []
