+++ src/pycropml/example.py
@@ -11,3 +11,4 @@ def example():
     parser = pp.Parser()
     model = parser.parse(fn)
     return model
+
