--- src/pycropml/test.py
@@ -1,8 +1,0 @@
-
-class Test(object):
-    """ Test """
-    def __init__(self, name, description, uri=None):
-        self.name = name
-        self.description = description
-        self.uri = uri
-        self.paramsets = []
